+++ network_module/timesync/roundtrip/roundtrip_subphase.h
@@ -39,7 +39,7 @@ public:
     unsigned long long getDuration() const override {
         return phaseDuration;
     }
-    unsigned long long getSlotsCount() const override { return 1; }
+    unsigned getSlotsCount() const override { return 1; }
     virtual ~RoundtripSubphase() {};
     
     static const unsigned int receiverWindow = 200000; //200us
